feat: Add Docker Desktop installation step
This commit adds the installation step for Docker Desktop to the `installer.sh` script. It downloads the .deb file from the official Docker website and installs the package using `apt`. Note that an error may be displayed during the installation process, but it can be ignored.

The purpose of this commit is to provide users with an easy way to install Docker Desktop on their Ubuntu systems.